Refactor asset preload and mission/card handling logic
Consolidated mission and card fetching logic into the preload assets hook to reduce redundancy and improve code organization. Removed unnecessary state and loaders from the `page.tsx` component in alignment with these changes. This improves maintainability and ensures centralized management of asset fetching.
