+++ utils/stores/frog-mahjong/useFrogMahjongStore.ts
@@ -218,7 +218,6 @@ const useFrogMahjongStore = create(
         clearMissionIDs: [],
         isRouletteLoading: false,
         isTimeOut: false,
-        allMissions: [],
         isHelpModalOpen: false,
         isUseItem: false,
       }),
